requirements.txt
streamlit
pyPDF2
google.generativeai
python-dotenv